test: move race tests to *_external_test.go files (#474)
Closes the file-naming gap identified by the retrospective
issue-closer audit on 2026-04-18. Issue #474 AC3 explicitly named:

  - webhook/webhook_external_test.go
  - file/file_external_test.go
  - syslog/syslog_external_test.go
  - loki/loki_external_test.go

PR #629 placed webhook correctly but put the other three in the
package's main *_test.go file. AC3 also specified `goleak.VerifyNone`
per-test; PR #629 relied on package-level `VerifyTestMain` — equivalent
in practice but not literally what AC3 required. This PR closes both.

Changes:

- Move TestFile_SetDiagnosticLoggerUnderEventLoad from
  file/file_test.go to file/file_external_test.go; unused `io` import
  removed from file_test.go.
- Move TestSyslog_SetDiagnosticLoggerUnderEventLoad from
  syslog/syslog_test.go to syslog/syslog_external_test.go.
- Move TestLoki_SetDiagnosticLoggerUnderEventLoad from
  loki/config_test.go to loki/loki_external_test.go; unused `io`,
  `sync`, `audit` imports removed from config_test.go.
- Add per-test goleak.VerifyNone(t) to all four tests, called AFTER
  the output's Close so background goroutines (rotate-mill, batch
  flush, accept loop) have exited before the assertion.
- Webhook test switched from `t.Cleanup(out.Close)` to explicit
  out.Close before the goleak call so the cleanup ordering is
  deterministic.

No production code change. Functional behaviour preserved — each
test still drives 100 concurrent iterations under -race.

AC5 (bench-compare regression check) is not addressed by this PR:
this is a test-reorganisation only, no hot-path change since #474
merged. Raising separately if re-verification is required.
